+++ lib/widgets/validate_photo.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 class Validate {
   final alert = new Alert();
 
-  validatePhoto(BuildContext context, int edad) {
+  validatePhoto(BuildContext context, double edad) {
     if (edad >= 0 && edad <= 18) {
       alert.showDialogMessage(context, Icons.error_outline, Colors.red,
           "Upps! Eres menor de edad :( ", edad, "¡no tienes acceso!");
